refactor(emit): split generic variadic tuple helpers
